requirements.txt
discord.py>=2.4.0
pytz
flask
gspread
oauth2client
pyasn1
